TODO: yes, if there are no houses, need to tear down all 3 hotels in this situation. as per the rules.
see:
https://en.wikibooks.org/wiki/Monopoly/Strategy

If there is a shortage of houses, players should consider buying the last houses rather than buying hotels because each hotel built returns four houses to the bank for other players to use. When there are no houses in the bank, no one else can build houses. Also, players with hotels can't easily convert them to cash in this situation; they would have to strip an entire colour group of its hotels to raise money since buildings must be removed evenly from a colour group and because the hotels cannot be exchanged for houses since there are no houses in the bank. The owners of inexpensive monopolies should be particularly eager to use this tactic to prevent the owners of expensive monopolies from fully developing.